// src/types/EventTypes.ts
export type EventTypeProps = {
    Eventname: string;
    description: string;
    date: string;
    time: string;
    duration: string;
    location: string;
    guests: string[];
    notification: string;
    remainder: string;
    file: any;

}


export type dataType = {
    Eventname: string;
    description: string;
    date: string;
    time: string;
    duration: string;
    location: string;
    guests: string[];
    notification: string;
    remainder: string;
    file: string;
}

export type EventType = {
    Eventname: string;
    date: string;
    time: string;
    duration: string;
    location: string;
    remainder: string;

}